# tests/STLCompliance.OpenApi.Tests/snapshots/routarr.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RoutArr.Api | v1"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/driver-portal/trips/{tripId}/accept": {
      "post": {
        "tags": [
          "DriverPortal"
        ],
        "operationId": "DriverPortalAcceptTrip",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/driver-portal/trips/{tripId}/capture-readiness": {
      "get": {
        "tags": [
          "DriverPortal"
        ],
        "operationId": "DriverPortalTripCaptureReadiness",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/driver-portal/trips/{tripId}/close": {
      "post": {
        "tags": [
          "DriverPortal"
        ],
        "operationId": "DriverPortalCloseTrip",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/driver-portal/trips/{tripId}/complete": {
      "post": {
        "tags": [
          "DriverPortal"
        ],
        "operationId": "DriverPortalCompleteTrip",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/driver-portal/trips/{tripId}/dispatch": {
      "post": {
        "tags": [
          "DriverPortal"
        ],
        "operationId": "DriverPortalDispatchTrip",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/driver-portal/trips/{tripId}/dvir": {
      "post": {
        "tags": [
          "DriverPortal"
        ],
        "operationId": "DriverPortalSubmitTripDvir",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTripDvirRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/driver-portal/trips/{tripId}/exceptions": {
      "post": {
        "tags": [
          "DriverPortal"
        ],
        "operationId": "DriverPortalReportTripException",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverPortalReportExceptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/driver-portal/trips/{tripId}/execution": {
      "get": {
        "tags": [
          "DriverPortal"
        ],
        "operationId": "DriverPortalTripExecution",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/driver-portal/trips/{tripId}/proofs": {
      "post": {
        "tags": [
          "DriverPortal"
        ],
        "operationId": "DriverPortalCreateTripProof",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripProofRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/driver-portal/trips/{tripId}/start": {
      "post": {
        "tags": [
          "DriverPortal"
        ],
        "operationId": "DriverPortalStartTrip",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/trips/{tripId}": {
      "get": {
        "tags": [
          "Trips"
        ],
        "operationId": "GetTrip",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/trips/{tripId}/capture-readiness": {
      "get": {
        "tags": [
          "TripExecutionCapture"
        ],
        "operationId": "GetTripCaptureReadiness",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/trips/{tripId}/dvir": {
      "get": {
        "tags": [
          "TripProofDvir"
        ],
        "operationId": "ListTripDvirInspections",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "TripProofDvir"
        ],
        "operationId": "SubmitTripDvir",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTripDvirRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/trips/{tripId}/execution": {
      "get": {
        "tags": [
          "TripProofDvir"
        ],
        "operationId": "GetTripExecutionSummary",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/trips/{tripId}/proofs": {
      "get": {
        "tags": [
          "TripProofDvir"
        ],
        "operationId": "ListTripProofs",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "TripProofDvir"
        ],
        "operationId": "CreateTripProof",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripProofRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/trips/{tripId}/capture-readiness": {
      "get": {
        "tags": [
          "TripExecutionCapture"
        ],
        "operationId": "GetTripCaptureReadinessV1",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTripProofRequest": {
        "required": [
          "proofType",
          "vehicleRefKey",
          "referenceKey",
          "notes",
          "capturedAt"
        ],
        "type": "object",
        "properties": {
          "proofType": {
            "type": "string"
          },
          "vehicleRefKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "referenceKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "capturedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "DriverPortalReportExceptionRequest": {
        "required": [
          "title",
          "description",
          "exceptionType"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "exceptionType": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SubmitTripDvirRequest": {
        "required": [
          "phase",
          "vehicleRefKey",
          "result",
          "odometerReading",
          "defectNotes"
        ],
        "type": "object",
        "properties": {
          "phase": {
            "type": "string"
          },
          "vehicleRefKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "result": {
            "type": "string"
          },
          "odometerReading": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "defectNotes": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}